[PRM-795] Updating routing for uploading additional documents

+++ app/controllers/OtherReasonController.scala
@@ -21,6 +21,11 @@ import config.{AppConfig, ErrorHandler}
 import controllers.predicates.{AuthPredicate, DataRequiredAction}
 import forms.WhenDidBecomeUnableForm
 import forms.WhyReturnSubmittedLateForm.whyReturnSubmittedLateForm
+import forms.upscan.{RemoveFileForm, UploadDocumentForm}
+import helpers.{FormProviderHelper, UpscanMessageHelper}
+import javax.inject.Inject
+import models.pages.{UploadAnotherDocumentPage, UploadFirstDocumentPage, WhenDidBecomeUnablePage, WhyWasReturnSubmittedLatePage, EvidencePage}
+import models.upload.UploadStatusEnum.READY
 import forms.upscan.{RemoveFileForm, UploadDocumentForm, YouHaveUploadedFilesForm}
 import helpers.{FormProviderHelper, UpscanMessageHelper}
 import models.pages.{EvidencePage, WhenDidBecomeUnablePage, WhyWasReturnSubmittedLatePage, YouHaveUploadedFilesPage}
@@ -41,8 +46,6 @@ import views.html.reasonableExcuseJourneys.other._
 import views.html.reasonableExcuseJourneys.other.noJs.{UploadAnotherDocumentPage, UploadFirstDocumentPage, UploadTakingLongerThanExpectedPage}
 import viewtils.RadioOptionHelper
 
-import java.time.LocalDate
-import javax.inject.Inject
 import scala.concurrent.{ExecutionContext, Future}
 
 class OtherReasonController @Inject()(whenDidBecomeUnablePage: WhenDidBecomeUnablePage,
@@ -126,14 +129,24 @@ class OtherReasonController @Inject()(whenDidBecomeUnablePage: WhenDidBecomeUnab
         for {
           previousUploadsState <- uploadJourneyRepository.getUploadsForJourney(request.session.get(SessionKeys.journeyId))
         } yield {
-          val previousUploads = previousUploadsState.fold("[]")(previousUploads => Json.stringify(Json.toJson(previousUploads)))
-          val initiateNextUploadUrl = controllers.routes.UpscanController.initiateCallToUpscan(request.session.get(SessionKeys.journeyId).get)
-          val getStatusUrl = controllers.routes.UpscanController.getStatusOfFileUpload(request.session.get(SessionKeys.journeyId).get, _)
-          val removeFileUrl = controllers.routes.UpscanController.removeFile(request.session.get(SessionKeys.journeyId).get, _)
-          val postAction = navigation.nextPage(EvidencePage, mode)
-          Ok(uploadEvidencePage(postAction, initiateNextUploadUrl, getStatusUrl, removeFileUrl, previousUploads))
+          if(request.cookies.get("jsenabled").isEmpty) {
+              mode match {
+                case NormalMode => Redirect(controllers.routes.OtherReasonController.onPageLoadForFirstFileUpload(mode))
+                case CheckMode => if (previousUploadsState.exists(_.count(_.fileStatus == READY) > 0)) {
+                  Redirect(controllers.routes.OtherReasonController.onPageLoadUploadList(mode))
+                } else {
+                  Redirect(controllers.routes.OtherReasonController.onPageLoadForFirstFileUpload(mode))
+                }
+              }
+          } else {
+            val previousUploads = previousUploadsState.fold("[]")(previousUploads => Json.stringify(Json.toJson(previousUploads)))
+            val initiateNextUploadUrl = controllers.routes.UpscanController.initiateCallToUpscan(request.session.get(SessionKeys.journeyId).get)
+            val getStatusUrl = controllers.routes.UpscanController.getStatusOfFileUpload(request.session.get(SessionKeys.journeyId).get, _)
+            val removeFileUrl = controllers.routes.UpscanController.removeFile(request.session.get(SessionKeys.journeyId).get, _)
+            val postAction = navigation.nextPage(EvidencePage, mode)
+            Ok(uploadEvidencePage(postAction, initiateNextUploadUrl, getStatusUrl, removeFileUrl, previousUploads))
+          }
         }
-      }
     }
   }
 
@@ -184,21 +197,28 @@ class OtherReasonController @Inject()(whenDidBecomeUnablePage: WhenDidBecomeUnab
           upscanResponseModel => {
             val optErrorCode: Option[String] = request.session.get(SessionKeys.errorCodeFromUpscan)
             val optFailureFromUpscan: Option[String] = request.session.get(SessionKeys.failureMessageFromUpscan)
+            val fileListPage = navigation.nextPage(UploadAnotherDocumentPage, mode)
             if(optErrorCode.isEmpty && optFailureFromUpscan.isEmpty) {
-              Ok(uploadAnotherDocumentPage(upscanResponseModel, formProvider))
+              Ok(uploadAnotherDocumentPage(upscanResponseModel, formProvider, fileListPage.url))
             } else if(optErrorCode.isDefined && optFailureFromUpscan.isEmpty) {
               val localisedFailureReason = UpscanMessageHelper.getUploadFailureMessage(optErrorCode.get)
               val formWithErrors = UploadDocumentForm.form.withError(FormError("file", localisedFailureReason))
-              BadRequest(uploadAnotherDocumentPage(upscanResponseModel, formWithErrors))
+              BadRequest(uploadAnotherDocumentPage(upscanResponseModel, formWithErrors, fileListPage.url))
                 .removingFromSession(SessionKeys.errorCodeFromUpscan)
             } else {
               val formWithErrors = UploadDocumentForm.form.withError(FormError("file", optFailureFromUpscan.get))
-              BadRequest(uploadAnotherDocumentPage(upscanResponseModel, formWithErrors))
+              BadRequest(uploadAnotherDocumentPage(upscanResponseModel, formWithErrors, fileListPage.url))
                 .removingFromSession(SessionKeys.failureMessageFromUpscan)
             }
           }
         )
       )
+  }
+
+  def onPageLoadUploadList(mode: Mode): Action[AnyContent] = (authorise andThen dataRequired) {
+    implicit request => {
+      Ok("uploaded file list")
+    }
   }
 
   def onPageLoadForUploadComplete(mode: Mode): Action[AnyContent] = (authorise andThen dataRequired).async {
@@ -212,8 +232,8 @@ class OtherReasonController @Inject()(whenDidBecomeUnablePage: WhenDidBecomeUnab
       val journeyId = request.session.get(SessionKeys.journeyId).get
       upscanService.getAmountOfFilesUploadedForJourney(journeyId).map(
         filesUploaded => {
-            Ok(youHaveUploadedFilesPage(formProvider, radioOptionsToRender, postAction, filesUploaded))
-         }
+          Ok(youHaveUploadedFilesPage(formProvider, radioOptionsToRender, postAction, filesUploaded))
+        }
       )
     }
   }
@@ -231,7 +251,7 @@ class OtherReasonController @Inject()(whenDidBecomeUnablePage: WhenDidBecomeUnab
                 BadRequest(youHaveUploadedFilesPage(formHasErrors, radioOptionsToRender, postAction, filesUploaded))
               } else {
                 Redirect(controllers.routes.OtherReasonController.onPageLoadForUploadComplete()) //TODO routing to be updated
-                }
+              }
             }
           )
         },
@@ -311,7 +331,7 @@ class OtherReasonController @Inject()(whenDidBecomeUnablePage: WhenDidBecomeUnab
                 .removingFromSession(SessionKeys.isAddingAnotherDocument, SessionKeys.fileReference))
             }
           } else {
-            Future(Redirect(controllers.routes.OtherReasonController.onPageLoadForUploadComplete())
+            Future(Redirect(controllers.routes.OtherReasonController.onPageLoadUploadList(mode))
               .removingFromSession(SessionKeys.isAddingAnotherDocument, SessionKeys.fileReference))
           }
         }
